requirements.txt
openai
google-genai
python-dotenv
pyautogui
pytesseract
pillow